integration-tests/requirements.txt
# Python dependencies for integration tests
numbersprotocol-capture-sdk==0.2.0
Pillow>=10.0.0
